More refactoring and extractions
