+++ server/netty-server/src/main/scala/sttp/tapir/server/netty/internal/StreamCompatible.scala
@@ -11,7 +11,11 @@ private[netty] trait StreamCompatible[S <: Streams[S]] {
   val streams: S
   def fromFile(file: FileRange): streams.BinaryStream
   def fromInputStream(is: () => InputStream, length: Option[Long]): streams.BinaryStream
+  def fromPublisher(publisher: Publisher[HttpContent], maxBytes: Option[Long]): streams.BinaryStream
   def asPublisher(s: streams.BinaryStream): Publisher[HttpContent]
+
+  def failedStream(e: => Throwable): streams.BinaryStream
+  def emptyStream: streams.BinaryStream
 
   def publisherFromFile(file: FileRange): Publisher[HttpContent] =
     asPublisher(fromFile(file))
